requirements.txt
Pillow
st-pages